Zmiana z int na float. Testy bytesToHuman
Na systemach 32 bitowych podczas liczenia długości pliku, int się przepełniał.

+++ app/Services/ChunkyUploader.php
@@ -74,7 +74,7 @@ class ChunkyUploader
         Storage::deleteDirectory(self::getChunkDirectory($request));
     }
 
-    public static function countDirectorySize($directory) : int
+    public static function countDirectorySize($directory) : float
     {
         $files = Storage::files($directory);
         $calculatedSize = 0;
